backend/django_microservices/requirements.txt
celery==5.3.5
Django>=4.1.6
django-extensions==3.2.3
djangorestframework==3.14.0
djongo>=1.3.6
dnspython==2.6.1
flower==2.0.1
pika==1.3.2
pymongo==3.12.3
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0

scikit-learn==1.3.2
joblib==1.3.2
numpy==1.24.4
python-dotenv==1.0.0
Sphinx==7.2.6
sphinx-rtd-theme==2.0.0

validators>=0.34.0
